// src/app/model/userModel.ts
export default class User{
    _id: string;
    username: string;
    firstname: String;
    lastname: String;
    email: String;
    password: String;
    isMember: String;
    organization: String;
    sector: String;
    address: String;
    contact_number: Number;
    email_alerts: [String];
    sms_alerts: [String];
}